:sparkles: Install plugin command
