+++ CONTRIBUTING.md
@@ -7,6 +7,7 @@ and [PHP-FIG Naming Conventions](https://github.com/php-fig/fig-standards/blob/m
 - Remember to follow [SemVer](http://semver.org). If you are changing the behavior,
 or the public api, you may need to update the docs.
 - Make sure that the current tests pass, and if you have added something new, add the tests where relevant.
+
 - Send a coherent commit history, making sure each individual commit in your pull request is meaningful. If you had to make multiple intermediate commits while developing, please [squash](http://git-scm.com/book/en/Git-Tools-Rewriting-History) them before submitting.
 - You may also need to [rebase](http://git-scm.com/book/en/Git-Branching-Rebasing) to avoid merge conflicts.
