+++ README.md
@@ -1,10 +1,10 @@
 
 
 ## Github Cms
-[https://github.com/juzaweb/laravel-cms](https://github.com/juzaweb/laravel-cms)
+[https://juzaweb.com/juzaweb/laravel-cms](https://github.com/juzaweb/laravel-cms)
 
 ## Documentation
-[Juzaweb cms documentation](https://juzaweb.com/cms)
+[Laravel cms](https://juzaweb.com/cms)
 
 ## Contributing
 Contributions are welcome, and are accepted via pull requests. Please review these guidelines before submitting any pull requests.
