+++ src/Console/Commands/Plugin/InstallCommand.php
@@ -3,10 +3,8 @@
 namespace Juzaweb\Console\Commands\Plugin;
 
 use Illuminate\Console\Command;
-use Juzaweb\Support\Json;
-use Juzaweb\Support\Process\Installer;
 use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputOption;
+use Juzaweb\Support\Manager\UpdateManager;
 
 class InstallCommand extends Command
 {
@@ -22,71 +20,39 @@ class InstallCommand extends Command
      *
      * @var string
      */
-    protected $description = 'Install the specified plugin by given package name (vendor/name).';
-
-    /**
-     * Create a new command instance.
-     */
-    public function __construct()
-    {
-        parent::__construct();
-    }
+    protected $description = 'Install plugin by name (vendor/name).';
 
     /**
      * Execute the console command.
      */
     public function handle()
     {
-        if (is_null($this->argument('name'))) {
-            $this->installFromFile();
-
-            return;
-        }
-
         $pluginName = $this->argument('name');
-        if (! file_exists($path = plugin_path($pluginName, 'composer.json'))) {
-            $this->error("File 'composer.json' does not exist in your project root.");
-
+        if (is_dir(config('juzaweb.plugin.path') . '/' . $pluginName)) {
+            $this->error("Plugin [{$pluginName}] already exist.");
             return;
         }
 
-        $modules = Json::make($path);
-        $dependencies = $modules->get('require', []);
+        $updater = new UpdateManager('plugin', $pluginName);
+        $this->info('Check file update');
+        $check = $updater->updateStep1();
 
-        foreach ($dependencies as $module) {
-            $module = collect($module);
+        if ($check) {
+            $this->info('Download File');
+            $updater->updateStep2();
 
-            $this->install(
-                $module->get('name'),
-                $module->get('version'),
-                $module->get('type')
-            );
-        }
-    }
+            $this->info('Unzip File');
+            $updater->updateStep3();
 
-    /**
-     * Install plugins from plugins.json file.
-     */
-    protected function installFromFile()
-    {
-        if (! file_exists($path = base_path('plugin.json'))) {
-            $this->error("File 'plugin.json' does not exist in your project root.");
+            $this->info('Move to folder');
+            $updater->updateStep4();
 
-            return;
-        }
+            $this->info('Update database');
+            $updater->updateStep5();
 
-        $modules = Json::make($path);
-
-        $dependencies = $modules->get('require', []);
-
-        foreach ($dependencies as $module) {
-            $module = collect($module);
-
-            $this->install(
-                $module->get('name'),
-                $module->get('version'),
-                $module->get('type')
-            );
+            $this->info('Plugin installation successful.');
+        } else {
+            $this->error("Plugin [{$pluginName}] does not exist.");
         }
     }
 
@@ -95,37 +61,11 @@ class InstallCommand extends Command
      *
      * @param string $name
      * @param string $version
-     * @param string $type
-     * @param bool   $tree
      */
-    protected function install($name, $version = 'dev-master', $type = 'composer', $tree = false)
+    protected function install($name, $version = 'dev-master')
     {
-        $installer = new Installer(
-            $name,
-            $version,
-            $type ?: $this->option('type'),
-            $tree ?: $this->option('tree')
-        );
-
-        $installer->setRepository($this->laravel['modules']);
-
-        $installer->setConsole($this);
-
-        if ($timeout = $this->option('timeout')) {
-            $installer->setTimeout($timeout);
-        }
-
-        if ($path = $this->option('path')) {
-            $installer->setPath($path);
-        }
-
-        $installer->run();
-
-        if (! $this->option('no-update')) {
-            $this->call('plugin:update', [
-                'module' => $installer->getModuleName(),
-            ]);
-        }
+        $updater = new UpdateManager('plugin', $name, $version);
+        $updater->update();
     }
 
     /**
@@ -136,24 +76,8 @@ class InstallCommand extends Command
     protected function getArguments()
     {
         return [
-            ['name', InputArgument::OPTIONAL, 'The name of plugin will be installed.'],
+            ['name', InputArgument::REQUIRED, 'The name of plugin will be installed.'],
             ['version', InputArgument::OPTIONAL, 'The version of plugin will be installed.'],
-        ];
-    }
-
-    /**
-     * Get the console command options.
-     *
-     * @return array
-     */
-    protected function getOptions()
-    {
-        return [
-            ['timeout', null, InputOption::VALUE_OPTIONAL, 'The process timeout.', null],
-            ['path', null, InputOption::VALUE_OPTIONAL, 'The installation path.', null],
-            ['type', null, InputOption::VALUE_OPTIONAL, 'The type of installation.', null],
-            ['tree', null, InputOption::VALUE_NONE, 'Install the plugin as a git subtree', null],
-            ['no-update', null, InputOption::VALUE_NONE, 'Disables the automatic update of the dependencies.', null],
         ];
     }
 }
