+++ src/Console/Commands/Plugin/InstallCommand2.php
@@ -1,0 +1,159 @@
+<?php
+
+namespace Juzaweb\Console\Commands\Plugin;
+
+use Illuminate\Console\Command;
+use Juzaweb\Support\Json;
+use Juzaweb\Support\Process\Installer;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputOption;
+
+class InstallCommand2 extends Command
+{
+    /**
+     * The console command name.
+     *
+     * @var string
+     */
+    protected $name = 'plugin:install';
+
+    /**
+     * The console command description.
+     *
+     * @var string
+     */
+    protected $description = 'Install the specified plugin by given package name (vendor/name).';
+
+    /**
+     * Create a new command instance.
+     */
+    public function __construct()
+    {
+        parent::__construct();
+    }
+
+    /**
+     * Execute the console command.
+     */
+    public function handle()
+    {
+        if (is_null($this->argument('name'))) {
+            $this->installFromFile();
+
+            return;
+        }
+
+        $pluginName = $this->argument('name');
+        if (! file_exists($path = plugin_path($pluginName, 'composer.json'))) {
+            $this->error("File 'composer.json' does not exist in your project root.");
+
+            return;
+        }
+
+        $modules = Json::make($path);
+        $dependencies = $modules->get('require', []);
+
+        foreach ($dependencies as $module) {
+            $module = collect($module);
+
+            $this->install(
+                $module->get('name'),
+                $module->get('version'),
+                $module->get('type')
+            );
+        }
+    }
+
+    /**
+     * Install plugins from plugins.json file.
+     */
+    protected function installFromFile()
+    {
+        if (! file_exists($path = base_path('plugin.json'))) {
+            $this->error("File 'plugin.json' does not exist in your project root.");
+
+            return;
+        }
+
+        $modules = Json::make($path);
+
+        $dependencies = $modules->get('require', []);
+
+        foreach ($dependencies as $module) {
+            $module = collect($module);
+
+            $this->install(
+                $module->get('name'),
+                $module->get('version'),
+                $module->get('type')
+            );
+        }
+    }
+
+    /**
+     * Install the specified plugin.
+     *
+     * @param string $name
+     * @param string $version
+     * @param string $type
+     * @param bool   $tree
+     */
+    protected function install($name, $version = 'dev-master', $type = 'composer', $tree = false)
+    {
+        $installer = new Installer(
+            $name,
+            $version,
+            $type ?: $this->option('type'),
+            $tree ?: $this->option('tree')
+        );
+
+        $installer->setRepository($this->laravel['modules']);
+
+        $installer->setConsole($this);
+
+        if ($timeout = $this->option('timeout')) {
+            $installer->setTimeout($timeout);
+        }
+
+        if ($path = $this->option('path')) {
+            $installer->setPath($path);
+        }
+
+        $installer->run();
+
+        if (! $this->option('no-update')) {
+            $this->call('plugin:update', [
+                'module' => $installer->getModuleName(),
+            ]);
+        }
+    }
+
+    /**
+     * Get the console command arguments.
+     *
+     * @return array
+     */
+    protected function getArguments()
+    {
+        return [
+            ['name', InputArgument::OPTIONAL, 'The name of plugin will be installed.'],
+            ['version', InputArgument::OPTIONAL, 'The version of plugin will be installed.'],
+        ];
+    }
+
+    /**
+     * Get the console command options.
+     *
+     * @return array
+     */
+    protected function getOptions()
+    {
+        return [
+            ['timeout', null, InputOption::VALUE_OPTIONAL, 'The process timeout.', null],
+            ['path', null, InputOption::VALUE_OPTIONAL, 'The installation path.', null],
+            ['type', null, InputOption::VALUE_OPTIONAL, 'The type of installation.', null],
+            ['tree', null, InputOption::VALUE_NONE, 'Install the plugin as a git subtree', null],
+            ['no-update', null, InputOption::VALUE_NONE, 'Disables the automatic update of the dependencies.', null],
+        ];
+    }
+}
