+++ src/Support/JuzawebApi.php
@@ -10,6 +10,8 @@
 
 namespace Juzaweb\Support;
 
+use GuzzleHttp\Exception\ClientException;
+
 class JuzawebApi
 {
     protected $curl;
@@ -109,19 +111,22 @@ class JuzawebApi
 
     protected function callApiGetData($method, $url, $params = [], $headers = [])
     {
-        switch (strtolower($method)) {
-            case 'post':
-                $response = $this->curl->post($url, $params, $headers);
-
-                break;
-            case 'put':
-                $response = $this->curl->put($url, $params, $headers);
-
-                break;
-            default:
-                $response = $this->curl->get($url, $params, $headers);
-
-                break;
+        try {
+            switch (strtolower($method)) {
+                case 'post':
+                    $response = $this->curl->post($url, $params, $headers);
+                    break;
+                case 'put':
+                    $response = $this->curl->put($url, $params, $headers);
+                    break;
+                default:
+                    $response = $this->curl->get($url, $params, $headers);
+                    break;
+            }
+        } catch (ClientException $e) {
+            $response = $e->getResponse();
+        } catch (\Throwable $e) {
+            throw $e;
         }
 
         $content = $response->getBody()->getContents();
