+++ src/resources/views/auth/forgot_password.blade.php
@@ -38,7 +38,6 @@
                     </div>
                 </div>
             </div>
-
         </div>
     </div>
 @endsection
